move tool statuses types to SDK
