+++ sdks/js/src/types.ts
@@ -684,6 +684,61 @@ export const WebsearchResultSchema = z.object({
 
 export type WebsearchResultPublicType = z.infer<typeof WebsearchResultSchema>;
 
+export const TOOL_EXECUTION_FINAL_STATUSES = [
+  "succeeded",
+  "errored",
+  "denied",
+] as const;
+
+export type ToolExecutionFinalStatus =
+  (typeof TOOL_EXECUTION_FINAL_STATUSES)[number];
+
+export const TOOL_EXECUTION_BLOCKED_STATUSES = [
+  "blocked_authentication_required",
+  "blocked_validation_required",
+] as const;
+
+export type ToolExecutionBlockedStatus =
+  (typeof TOOL_EXECUTION_BLOCKED_STATUSES)[number];
+
+export const TOOL_EXECUTION_TRANSIENT_STATUSES = [
+  "ready_allowed_explicitly",
+  "ready_allowed_implicitly",
+  ...TOOL_EXECUTION_BLOCKED_STATUSES,
+  "running",
+] as const;
+
+export type ToolExecutionTransientStatus =
+  (typeof TOOL_EXECUTION_TRANSIENT_STATUSES)[number];
+
+export type ToolExecutionStatus =
+  | ToolExecutionFinalStatus
+  | ToolExecutionTransientStatus;
+
+export function isToolExecutionStatusFinal(
+  state: ToolExecutionStatus
+): state is ToolExecutionFinalStatus {
+  return TOOL_EXECUTION_FINAL_STATUSES.includes(
+    state as ToolExecutionFinalStatus
+  );
+}
+
+export function isToolExecutionStatusTransient(
+  state: ToolExecutionStatus
+): state is ToolExecutionTransientStatus {
+  return TOOL_EXECUTION_TRANSIENT_STATUSES.includes(
+    state as ToolExecutionTransientStatus
+  );
+}
+
+export function isToolExecutionStatusBlocked(
+  state: ToolExecutionStatus
+): state is ToolExecutionBlockedStatus {
+  return TOOL_EXECUTION_BLOCKED_STATUSES.includes(
+    state as ToolExecutionBlockedStatus
+  );
+}
+
 const MCPActionTypeSchema = z.object({
   id: ModelIdSchema,
   mcpServerId: z.string().nullable(),
@@ -692,6 +747,10 @@ const MCPActionTypeSchema = z.object({
   functionCallName: z.string().nullable(),
   params: z.record(z.any()),
   output: CallToolResultSchema.shape.content.nullable(),
+  status: z.enum([
+    ...TOOL_EXECUTION_FINAL_STATUSES,
+    ...TOOL_EXECUTION_TRANSIENT_STATUSES,
+  ]),
   type: z.literal("tool_action"),
 });
